+++ crates/holochain/src/conductor.rs
@@ -30,4 +30,6 @@ pub use cell::Cell;
 pub use conductor::Conductor;
 pub use conductor::ConductorBuilder;
 pub use conductor::ConductorHandle;
+pub use conductor::ConductorNetworkState;
+pub use conductor::NetworkEvent;
 pub use conductor::{full_integration_dump, integration_dump};